+++ examples/copyContentToProjectFolder/applications/example/app/examples/mysql/getCities.json/get.js
@@ -1,6 +1,6 @@
 module.exports = function(client, callback) {
 
-	dbAlias.query('select * from City', function(err, rows, fields) {
+	aliasNameMy.query('select * from City', function(err, rows, fields) {
 		callback({ rows:rows, fields:fields });
 	});