+++ examples/copyContentToProjectFolder/applications/example/app/examples/tools/serverHealth.json/get.js
@@ -1,8 +1,9 @@
 module.exports = function(client, callback) {
 
 	application.leak = application.leak || [];
-	var n = application.leak.length ? application.leak.length : 0;
-	application.leak.push(new Array(1024*(1024/n+1)).join("Hello node.js, Hello Impress Application Server "));
+	var n = client.application.leak.length;
+	n = 1024*Math.round(1024);
+	application.leak.push(new Array(n).join("Hello node.js, Hello Impress Application Server "));
 
 	callback({
 		memory: process.memoryUsage()