+++ HISTORY.md
@@ -1,3 +1,11 @@
+0.1.40 / 2014-06-10
+==================
+
+  * Implemented log files separation for applications, see: https://github.com/tshemsedinov/impress/issues/47
+  * Refactored /lib/impress.log.js
+  * Added wcl.js to example application for accessing server-side APIs (predefined and introspective)
+  * Fixed impress.security.js
+
 0.1.39 / 2014-06-08
 ==================